+++ NEWS.md
@@ -8,6 +8,7 @@
 - **BREAKING CHANGE** - removed `delay` parameter from `hide`/`show`/`toggle` because the same behaviour can now be achieved using the `delay` function (if you previously used `show(id = "text", delay = 2)`, you can now use `delay(2000, show("text"))`)
 - added `onevent` function, which is similar to `onclick` but can be used for many other different events
 - `reset` now works with dynamically generated elements (inputs that are created with `renderUI`)
+- make `disabled` work for dynamically generated elements
 - removed defunc `resettable` function
 
 # shinyjs 0.1.1